forthewy/shopForthewy_221103#11 아이템 테스트 중

+++ src/main/java/com/shoppingmall/item/ItemRestController.java
@@ -5,10 +5,12 @@ import java.util.Map;
 
 import javax.servlet.http.HttpSession;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.util.ObjectUtils;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -20,6 +22,8 @@ import com.shoppingmall.item.bo.ItemBO;
 @RestController
 public class ItemRestController {
 
+	private Logger log = LoggerFactory.getLogger(this.getClass());
+	
 	@Autowired
 	private ItemBO itemBO;
 	
@@ -50,6 +54,13 @@ public class ItemRestController {
 		Map<String, Object> result = new HashMap<>();
 		
 		String sellerLoginId = (String) session.getAttribute("userLoginId");
+		// 비로그인이라면 
+		if (ObjectUtils.isEmpty(sellerLoginId)) {
+			result.put("code", 550);
+			result.put("errorMessage", "로그인되어있지 않습니다");
+			return result;
+		}		
+		
 		
 		// DB insert
 		int row = itemBO.addItem(sellerLoginId, name, number, price, content, sort, thumbnailImg, deliveryPrice);
@@ -60,6 +71,7 @@ public class ItemRestController {
 		} else {
 			result.put("code", 500);
 			result.put("result", "상품 등록에 실패했습니다.");
+			log.error("[상품 등록] 상품 등록 실패 sellerLoginId:{}", sellerLoginId);
 		}
 		
 		return result;
@@ -96,15 +108,26 @@ public class ItemRestController {
 		
 		String sellerLoginId = (String) session.getAttribute("userLoginId");
 		
+		// 비로그인이라면 
+		if (ObjectUtils.isEmpty(sellerLoginId)) {
+			result.put("code", 550);
+			result.put("errorMessage", "로그인되어있지 않습니다");
+			return result;
+		}
+		
 		// DB update
 		int row = itemBO.updateItem(sellerLoginId, name, number, price, content, sort, thumbnailImg, deliveryPrice, itemId);
 		
 		if (row > 0) {
 			result.put("code", 300);
 			result.put("result", "success");
+			log.info("[상품 수정] 상품 정보 수정 성공 name:{}, number:{}, price:{}, content:{}, sort:{}, thumbnailImg:{},"
+					+ "deliveryPrice:{}, itemId:{}", name, number, price, content, sort, thumbnailImg, deliveryPrice, itemId);
 		} else {
 			result.put("code", 500);
 			result.put("errorMessage", "상품 수정에 실패했습니다.");
+			log.error("[상품 수정] 상품 정보 수정 실패 name:{}, number:{}, price:{}, content:{}, sort:{}, thumbnailImg:{},\"\r\n"
+					+ "deliveryPrice:{}, itemId:{}, sellerLoginId:{}",  name, number, price, content, sort, thumbnailImg, deliveryPrice, itemId, sellerLoginId);
 		}
 		
 		return result;
@@ -123,15 +146,27 @@ public class ItemRestController {
 		
 		Map<String, Object> result = new HashMap<>();
 		
+		Integer userId = (Integer) session.getAttribute("userId");
+		
+		// 비로그인이라면 
+		if (ObjectUtils.isEmpty(userId)) {
+			result.put("code", 550);
+			result.put("errorMessage", "로그인되어있지 않습니다");
+			return result;
+		}
+		
+		
 		//DB delete
 		int row = itemBO.deleteItem(itemId);
 				
 		if (row > 0) {
 			result.put("code", 300);
 			result.put("result", "success");
+			log.info("[상품 삭제] 상품 삭제 성공 itemId:{}", itemId); 
 		} else {
 			result.put("code", 500);
 			result.put("errorMessage", "상품 삭제에 실패했습니다.");
+			log.error("[상품 삭제] 상품 삭제 실패 itemId:{}", itemId); 
 		}
 				
 		return result;
